分层

+++ app/admin/dao/rbac.go
@@ -1,30 +1,36 @@
-package database
+package dao
 
 import (
-	"ErotsServer/app/admin/structure"
+	userDao "ErotsServer/app/user/dao"
 
 	"github.com/ipuppet/gtools/database"
 )
 
 func GetRoles() ([]map[string]interface{}, error) {
-	return database.SQLQueryRetrieveMap(Db, "select * from ums.rbac_role")
+	return database.SQLQueryRetrieveMapNoCache(Db, "select * from ums.rbac_role")
 }
 
-func UpdateRole(role structure.Role) error {
+func UpdateRole(role userDao.Role) error {
 	_, err := database.MustExec(Db.Exec(`update ums.rbac_role
 		set role_id=?, name=?, description=?
 		where role_id=?`,
-		*role.RoleId, role.Name, role.Description, *role.RoleId))
+		role.RoleId, role.Name, role.Description, role.RoleId))
+	if err != nil {
+		Logger.Println(err)
+	}
 
 	return err
 }
 
-func AddRole(role structure.Role) error {
+func AddRole(role userDao.Role) error {
 	_, err := database.MustExec(Db.Exec(`insert into ums.rbac_role
 		(role_id, name, description)
 		values
 		(?, ?, ?)`,
-		*role.RoleId, role.Name, role.Description))
+		role.RoleId, role.Name, role.Description))
+	if err != nil {
+		Logger.Println(err)
+	}
 
 	return err
 }
@@ -39,13 +45,17 @@ func DeleteRole(roleId int) error {
 	_, err = database.MustExec(conn.Exec("delete from ums.rbac_role where role_id=?", roleId))
 	if err != nil {
 		conn.Rollback()
+		Logger.Println(err)
 		return err
 	}
 
-	_, err = database.MustExec(conn.Exec("delete from ums.rbac_role_permission where role_id=?", roleId))
-	if err != nil {
-		conn.Rollback()
-		return err
+	if err := conn.QueryRow("select count(permission_id) as count from ums.rbac_role_permission").Scan(); err == nil {
+		_, err = database.MustExec(conn.Exec("delete from ums.rbac_role_permission where role_id=?", roleId))
+		if err != nil {
+			conn.Rollback()
+			Logger.Println(err)
+			return err
+		}
 	}
 
 	conn.Commit()
@@ -54,24 +64,30 @@ func DeleteRole(roleId int) error {
 }
 
 func GetPermissions() ([]map[string]interface{}, error) {
-	return database.SQLQueryRetrieveMap(Db, "select * from ums.rbac_permission")
+	return database.SQLQueryRetrieveMapNoCache(Db, "select * from ums.rbac_permission")
 }
 
-func UpdatePermission(permission structure.Permission) error {
+func UpdatePermission(permission userDao.Permission) error {
 	_, err := database.MustExec(Db.Exec(`update ums.rbac_permission
 		set permission_id=?, module=?, description=?
 		where permission_id=?`,
-		*permission.PermissionId, permission.Name, permission.Description, *permission.PermissionId))
+		permission.PermissionId, permission.Module, permission.Description, permission.PermissionId))
+	if err != nil {
+		Logger.Println(err)
+	}
 
 	return err
 }
 
-func AddPermission(permission structure.Permission) error {
+func AddPermission(permission userDao.Permission) error {
 	_, err := database.MustExec(Db.Exec(`insert into ums.rbac_permission
 		(permission_id, module, description)
 		values
 		(?, ?, ?)`,
-		*permission.PermissionId, permission.Name, permission.Description))
+		permission.PermissionId, permission.Module, permission.Description))
+	if err != nil {
+		Logger.Println(err)
+	}
 
 	return err
 }
@@ -86,13 +102,17 @@ func DeletePermission(permissionId int) error {
 	_, err = database.MustExec(conn.Exec("delete from ums.rbac_permission where permission_id=?", permissionId))
 	if err != nil {
 		conn.Rollback()
+		Logger.Println(err)
 		return err
 	}
 
-	_, err = database.MustExec(conn.Exec("delete from ums.rbac_role_permission where permission_id=?", permissionId))
-	if err != nil {
-		conn.Rollback()
-		return err
+	if err := conn.QueryRow("select count(permission_id) as count from ums.rbac_role_permission").Scan(); err == nil {
+		_, err = database.MustExec(conn.Exec("delete from ums.rbac_role_permission where permission_id=?", permissionId))
+		if err != nil {
+			conn.Rollback()
+			Logger.Println(err)
+			return err
+		}
 	}
 
 	conn.Commit()
@@ -101,7 +121,7 @@ func DeletePermission(permissionId int) error {
 }
 
 func GetRolePermissions(roleId int) ([]map[string]interface{}, error) {
-	return database.SQLQueryRetrieveMap(Db,
+	return database.SQLQueryRetrieveMapNoCache(Db,
 		`select a.permission_id,b.module,b.description
 		from ums.rbac_role_permission a
 		left join ums.rbac_permission b on a.role_id=?
@@ -112,6 +132,9 @@ func GetRolePermissions(roleId int) ([]map[string]interface{}, error) {
 func DeleteRolePermission(roleId int, permissionId int) error {
 	_, err := database.MustExec(Db.Exec(`delete from ums.rbac_role_permission where role_id=? and permission_id=?`,
 		roleId, permissionId))
+	if err != nil {
+		Logger.Println(err)
+	}
 
 	return err
 }
@@ -122,6 +145,9 @@ func AddRolePermission(roleId int, permissionId int) error {
 		values
 		(?, ?)`,
 		roleId, permissionId))
+	if err != nil {
+		Logger.Println(err)
+	}
 
 	return err
 }
